+++ index.d.ts
@@ -92,7 +92,7 @@ export namespace StorageChannelEventType {
 export class Ownership<K extends string> {
   constructor(name: K);
   take(key: K, age: number): boolean;
-  take(key: K, age: number, wait: number): AtomicPromise<void>;
+  take(key: K, age: number, wait: number): Promise<boolean>;
   extend(key: K, age: number): boolean;
   release(key: K): void;
   close(): void;